Some Quick Fixes
Did some fixes I could do quickly, namely all the GFs and BFs that weren't working (except for Imposter mod because the sprites for the Beans perplex and confuse me (still broken)).  Haven't fixed Kapi yet because lazy.  Will probably do Kapi next.  Also, I was gonna quickly add Hellclown Tricky so that when I do stages I can just reduce the camZoom, but did you know that each Anim for Phase 3 Tricky is a SEPERATE FILE!?!?!?  WTF WHY IS IT LIKE THAT.  this means that I have to either A. reimplement a system in VoidEngine that allows sprites to be loaded like this, or B. get the fla and edit it in animate to fucking get it working properly.  Fuck Tricky's clown ass